+++ ejercicioinn13/ejercicioinn13/src/main/java/com/example/ejercicioinn13/repository/DepartamentoRepository.java
@@ -3,5 +3,5 @@ package com.example.ejercicioinn13.repository;
 import com.example.ejercicioinn13.entity.Departamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartamentoRepository extends JpaRepository<Departamento,Integer> {
+public interface DepartamentoRepository extends JpaRepository<Departamento,Long> {
 }
